refactor: move heirs services into heirs services
