+++ frontend/src/components/Modal/HeirsModal/HeirsPersonModal/HeirsPersonModal.tsx
@@ -3,7 +3,7 @@ import { Form, Formik } from 'formik'
 import { ObjectSchema, array, mixed, object, string } from 'yup'
 import { personMoreInfosOptions } from '../../../../../content/checkboxOptions'
 import { childRelationshipOptions, genderOptions, heirsPersonType } from '../../../../../content/dropdownOptions'
-import { heirsTypes } from '../../../../app/(dynamic)/last-will/editor/heirs/heirs'
+import { heirsTypes } from '../../../../services/heirs'
 import { useAppDispatch, useAppSelector } from '../../../../store/hooks'
 import { addPersonHeir, sendLastWillState, updatePersonHeir } from '../../../../store/lastwill'
 import { Gender } from '../../../../types/gender'
